debug mutex

+++ my-async-framework/sync/atomic.hpp
@@ -5,6 +5,6 @@
 namespace MyAsyncFramework::sync {
 
 // TODO: add custom atomic using asm (and be able to stay both macOs/Linux compatible)
-using Atomic = typename std::atomic<int64_t>;
+using Atomic = typename std::atomic<uint32_t>;
 
 }
